Copy ID to clipboard and cross-linking progress

+++ data/data-action.php
@@ -53,17 +53,24 @@ function projectRemove() {
 	if($_POST["project"] === "all") {
 		$jsons = glob("project/json/*.json", GLOB_BRACE);
 		$pngs = glob("project/img/*.png", GLOB_BRACE);
+		foreach($jsons as $json) {
+			unlink($json);
+		}
+		foreach($pngs as $png) {
+			unlink($png);
+		}
 	} else {
 		$png = 'project/img/' . $_POST["project"] . '.png';
 		$json = 'project/json/' . $_POST["project"] . '.json';
+		if(file_exists($png)) {
+			unlink($png);
+		}
+		if(file_exists($json)) {
+			unlink($json);
+		}
 	}
 
-	foreach($jsons as $json) {
-		unlink($json);
-	}
-	foreach($pngs as $png) {
-		unlink($png);
-	}
+
 	removeDirectory("project/web");
 }
 
@@ -71,23 +78,16 @@ function staffAction() {
 	if($_FILES["staff_img"]["tmp_name"] != null) {
 		imagepng(imagecreatefromstring(file_get_contents($_FILES["staff_img"]["tmp_name"])), 'staff/img/' . $_POST["filename"] . '.png');
 	}
-
 	$fn = 'staff/json/' . $_POST["filename"] . '.json';
-	$linked = array();
-	if(!isset($_POST["linked"]) || in_array("none", $_POST["linked"])) {
-		$linked = array("");
-	} else {
-		$linked = $_POST["linked"];
-	}
-	$data = array($_POST["staff_name"], MarkdownExtended::parse($_POST["staff_desc"])->getContent(), $_POST["staff_desc"], $linked);
+	$data = array($_POST["staff_name"], MarkdownExtended::parse($_POST["staff_desc"])->getContent(), $_POST["staff_desc"]);
 	ftruncate(fopen($fn, "r+"), 0);
-	file_put_contents($fn, json_encode($data, FILE_APPEND));
+	file_put_contents($fn, json_encode($data), FILE_APPEND);
 }
 
 function staffAdd() {
 	include("../utils/phputils.php");
 	$filename = 'staff/json/' . generateRandomString(8) . '.json';
-	file_put_contents($filename, json_encode(array("","","",array(""))));
+	file_put_contents($filename, json_encode(array("","","")));
 	chmod($filename, 0755);
 }
 
@@ -114,5 +114,5 @@ function staffRemove() {
 }
 
 echo '<META HTTP-EQUIV="CACHE-CONTROL" CONTENT="NO-CACHE">';
-header("refresh:10; url=../admin/");
+header("refresh:0; url=../admin/");
 ?>
